Add DependencyInjection classes for Application and Infrastructure
Two new static classes named `DependencyInjection` have been added to the `BlazorShop.Application` and `BlazorShop.Infrastructure` namespaces.

In the `BlazorShop.Application` namespace, the `DependencyInjection` class includes a method `AddApplication` that extends `IServiceCollection` to return the service collection.

In the `BlazorShop.Infrastructure` namespace, the `DependencyInjection` class includes a method `AddInfrastructure` that extends `IServiceCollection` to add a scoped service for `IGenericRepository<>` with an implementation of `GenericRepository<>`, and then returns the service collection.