Add loading screen and apply background style to body
Modified the `<body>` tag in `index.html` to include a `class`
attribute with the value `site-bg` for applying a background
style. Added a `#app` container with a `loader_main` div to
implement a loading screen. The loading screen includes a
spinner (`spinner` class) with an `aria-label` for accessibility
and a `loading-progress-text` div for displaying progress text.